% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cprobit.R
\name{compile_est}
\alias{compile_est}
\title{Inpernal function: generate commonly used summary statistics for estimates.}
\usage{
compile_est(
  var,
  est,
  se = NULL,
  z_score = NULL,
  pval = NULL,
  value_null = 0,
  ci_lower = NULL,
  ci_upper = NULL,
  prefix = NULL,
  postfix = NULL
)
}
\arguments{
\item{var}{Names of variables.}

\item{est}{Estimated regression coefficients.}

\item{se}{SE of estimates.}

\item{z_score}{Z score of estimates, i.e., \code{est / se}.}

\item{pval}{P-value of estimates.}

\item{value_null}{Null effects for estimates, either with length 1 or length
of \code{est}. Default is 0.}

\item{ci_lower}{Lower bound of 95\% CI of estimates.}

\item{ci_upper}{Upper bound of 95\% CI of estimates.}

\item{prefix}{Prefix to the column names in the \code{data.frame} returned.}

\item{postfix}{Postfix to the column names in the \code{data.frame} returned.}
}
\description{
Inpernal function: generate commonly used summary statistics for estimates.
}
\details{
Vectorised, as long as the length of the input match.
}
