% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cprobit.R
\name{cprobit_step1}
\alias{cprobit_step1}
\title{Inpernal function: step 1 of the proposed workflow}
\usage{
cprobit_step1(y_name, x_names, dat_diff, var_names = NULL)
}
\arguments{
\item{y_name}{Name of outcome variable for Step 1.}

\item{x_names}{Names of covariates for Step 1.}

\item{dat_diff}{A \code{data.frame} containing the difference data.}

\item{var_names}{Variable names for the estimates.}
}
\value{
Returns a \code{data.frame} summarising the Step 1 estimates
(\code{coef}) and the covariance matrix for the Step 1 estimates (\code{vcov}).
}
\description{
Implements the Step 1 of the proposed workflow, where a cprobit
  model is applied to analyse whether there is an increase in the outcome
  within each subject.
}
