% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cprobit.R
\name{geom_mean}
\alias{geom_mean}
\title{Inpernal function: compute geometric mean of a positive variable}
\usage{
geom_mean(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
Inpernal function: compute geometric mean of a positive variable
}
