% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cprobit.R
\name{get_v}
\alias{get_v}
\title{Inpernal function: compute difference in the (transformed) outcome}
\usage{
get_v(y1, y2, lambda = NA, scaled = TRUE)
}
\arguments{
\item{y1}{Numeric vector of the observed outcome at observation time 1.}

\item{y2}{Numeric vector of the observed outcome at observation time 2.}

\item{lambda}{The Box-Cox transformation parameter. Default is \code{NA},
indicating no need for a transformation. See \code{Details}.}

\item{scaled}{Whether the difference in the transformed outomes should be
scaled by the Jacobian.}
}
\value{
Returns the difference in the observed outcomes if \code{lambda =
  NA}, or the difference in the scaled transformed outcomes with
  transformation parameter \code{lambda}.
}
\description{
Inpernal function: compute difference in the (transformed) outcome
}
