% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cprobit.R
\name{make_design_mat}
\alias{make_design_mat}
\title{Inpernal function: construct design matrix without the intercept term.}
\usage{
make_design_mat(lp, dat, remove_intercept = TRUE)
}
\arguments{
\item{lp}{Formula for the linear predictor part, as a string.}

\item{dat}{Data to construct the design matrix from.}

\item{remove_intercept}{Whether the first column should be removed. Default
is \code{TRUE} (to remove the intercept term).}
}
\value{
Returns a list containing the constructed design matrix and the
  original variable names. In the column names of the design matrix returned
  , any \code{:} in variable names are replaced with \code{.} to avoid
  computational issues when using the design matrix to fit model.
}
\description{
Inpernal function: construct design matrix without the intercept term.
}
