% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cprobit.R
\name{update_estimate}
\alias{update_estimate}
\title{Inpernal function: update Step 1 estimates to obtain linear exposure effect on
(transformed) outcome}
\usage{
update_estimate(
  y1_name,
  y2_name,
  var_names = NULL,
  dat_diff,
  res_step1,
  transform = FALSE
)
}
\arguments{
\item{y1_name}{Name of observed outcome at observation time 1.}

\item{y2_name}{Name of observed outcome at observation time 2.}

\item{var_names}{Variable names for the estimates.}

\item{dat_diff}{A \code{data.frame} containing the difference data.}

\item{res_step1}{Results from Step 1 of the workflow.}

\item{transform}{Whether the outcome should be transformed. Default is
\code{FALSE}.}
}
\value{
Returns a \code{list}: a \code{data.frame} summarising the estimated
  linear exposure effect, the estimated standard deviation of the error terms
  from the difference model, the covariance matrix of the estimated exposure
  effects, a \code{data.frame} summarising the estimated transforamtion
  parameter, and the residuals.
}
\description{
Inpernal function: update Step 1 estimates to obtain linear exposure effect on
(transformed) outcome
}
